requirements.txt
nextcord
schedule
dateparser
mysql-connector-python
json5
pandas
pillow
apscheduler